Save macro calculation results to database

+++ backend/MareksGym.Api/Application/Macros/MacroCalculationInput.cs
@@ -3,7 +3,7 @@
 public sealed record MacroCalculationInput(
     Sex Sex,
     int Age,
-    double HeightCm,
+    int HeightCm,
     double WeightKg,
     ActivityLevel ActivityLevel,
     Goal Goal
