+++ backend/MareksGym.Api/Contracts/Macros/CalculateMacrosRequest.cs
@@ -3,7 +3,7 @@
 public sealed record CalculateMacrosRequest(
     string Sex,     // "male" | "female"
     int Age,
-    double HeightCm,
+    int HeightCm,
     double WeightKg,
     string ActivityLevel,   // "sedentary" | "light" | "moderate" | "active" | "very_active"
     string Goal     // "cut" | "maintain" | "bulk"
